+++ csharp/Svix.Tests/WebhookTest.cs
@@ -4,7 +4,6 @@ using System;
 using System.Net;
 
 using Svix.Exceptions;
-using Svix.Models;
 
 namespace Svix.Tests
 {
